portal/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
google-cloud-firestore==2.16.1
python-multipart==0.0.9
